arch: arquitetura alterada para separar a biblioteca do arquivo principal
