+++ lib/batalha_naval.h
@@ -1,0 +1,5 @@
+int tabuleiro(int linha, char coluna, int linha_bot, int coluna_bot);
+char verifica_casa(int i, int j, int player);
+int numero_coluna(char coluna);
+int posiciona_navio();
+int verifica_vitoria();
